requirements.txt
Flask>=3.0.0
qrcode>=7.4.0
Pillow>=10.0.0
